Moves the system program into its own crate in "programs/system".
