+++ Cargo.toml
@@ -61,6 +61,7 @@ members = [
     "programs/ed25519-tests",
     "programs/loader-v3",
     "programs/stake",
+    "programs/system",
     "programs/vote",
     "programs/zk-token-proof",
     "programs/zk-token-proof-tests",
@@ -348,6 +349,7 @@ solana-storage-bigtable = { path = "storage-bigtable", version = "=1.16.0" }
 solana-storage-proto = { path = "storage-proto", version = "=1.16.0" }
 solana-streamer = { path = "streamer", version = "=1.16.0" }
 solana-sys-tuner = { path = "sys-tuner", version = "=1.16.0" }
+solana-system-program = { path = "programs/system", version = "=1.16.0" }
 solana-test-validator = { path = "test-validator", version = "=1.16.0" }
 solana-thin-client = { path = "thin-client", version = "=1.16.0" }
 solana-tpu-client = { path = "tpu-client", version = "=1.16.0", default-features = false }
