+++ programs/system/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "solana-system-program"
+description = "Solana System program"
+documentation = "https://docs.rs/solana-system-program"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+bincode = { workspace = true }
+log = { workspace = true }
+serde = { workspace = true }
+serde_derive = { workspace = true }
+solana-program-runtime = { workspace = true }
+solana-sdk = { workspace = true }
+
+[dev-dependencies]
+assert_matches = { workspace = true }
+solana-logger = { workspace = true }
+solana-runtime = { workspace = true }
+
+[lib]
+crate-type = ["lib"]
+name = "solana_system_program"
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
