+++ runtime/src/lib.rs
@@ -48,7 +48,6 @@ pub mod inline_spl_token_2022;
 pub mod loader_utils;
 pub mod message_processor;
 pub mod non_circulating_supply;
-mod nonce_keyed_account;
 pub mod prioritization_fee;
 pub mod prioritization_fee_cache;
 mod pubkey_bins;
@@ -76,7 +75,6 @@ pub mod stakes;
 pub mod static_ids;
 pub mod status_cache;
 mod storable_accounts;
-mod system_instruction_processor;
 pub mod transaction_batch;
 pub mod transaction_error_metrics;
 pub mod transaction_priority_details;
